Make preventing default on keydown optional
Preventing default on keydown causes all inputs on the page to stop working whilst the Menu component is mounted. Some apps have the menu mounted whilst not in use, causing inputs to stop working.

Adds a flag to disable this feature and also makes it work as intended only for keys required for operating the menu that may cause the window to scroll.

Related issue https://github.com/fkhadra/react-contexify/issues/168